Move writeobject into its own file
